MEAクラスからグラフ描画部分をFigMEAクラスに切り出す
