+++ pyMEA/__init__.py
@@ -1,5 +1,6 @@
 from pyMEA.calculator.calculator import Calculator
 from pyMEA.CardioAveWave import CardioAveWave
+from pyMEA.figure.FigMEA import FigMEA
 from pyMEA.FilterMEA import FilterMEA
 from pyMEA.find_peaks.peak_detection import detect_peak_neg, detect_peak_pos
 from pyMEA.fit_gradient import calc_gradient_velocity
@@ -7,6 +8,7 @@ from pyMEA.MEA import MEA
 
 __all__ = [
     "MEA",
+    "FigMEA",
     "Calculator",
     "FilterMEA",
     "CardioAveWave",
