+++ pyMEA/figure/plot/histogram.py
