+++ pyMEA/figure/plot/raster_plot.py
@@ -13,7 +13,7 @@ def raster_plot(
     figsize=(8, 8),
     start=0,
     end=120,
-    dpi=300
+    dpi=300,
 ) -> None:
     plt.figure(figsize=figsize, dpi=dpi)
     for i, ele in enumerate(eles):
