+++ test/test_calculator.py
@@ -5,6 +5,7 @@ import numpy as np
 
 from pyMEA import detect_peak_neg
 from pyMEA.calculator.calculator import Calculator
+from pyMEA.figure.FigMEA import FigMEA
 from pyMEA.MEA import MEA
 
 
@@ -15,6 +16,7 @@ class CalculatorTest(unittest.TestCase):
         self.peak_index = detect_peak_neg(self.data.array)
         self.calc450 = Calculator(self.data, 450)
         self.calc150 = Calculator(self.data, 150)
+        self.fm = FigMEA(self.data)
 
     def test_ISIが正しく計算できる(self):
         isi = self.calc450.isi(self.peak_index, 32)
@@ -58,7 +60,7 @@ class CalculatorTest(unittest.TestCase):
 
     @patch("matplotlib.pyplot.show")
     def test_速度ベクトルから伝導速度が正しく計算できる(self, mock_show: MagicMock):
-        popts, r2s = self.data.draw_2d(self.peak_index, 450)
+        popts, r2s = self.fm.draw_2d(self.peak_index, 450)
         cvs = self.calc450.gradient_velocity(popts)
 
         for cv in cvs:
