+++ test/test_figure.py
@@ -3,6 +3,7 @@ from unittest.mock import MagicMock, patch
 
 import numpy as np
 
+from pyMEA.figure.FigMEA import FigMEA
 from pyMEA.find_peaks.peak_detection import detect_peak_neg
 from pyMEA.MEA import MEA
 
@@ -12,11 +13,12 @@ class MyTestCase(unittest.TestCase):
         self.path = "./public/230615_day2_test_5s_.hed"
         self.data = MEA(self.path, 0, 5)
         self.peak_index = detect_peak_neg(self.data.array)
+        self.fm = FigMEA(self.data)
 
     @patch("matplotlib.pyplot.show")
     @patch("matplotlib.pyplot.plot")
     def test_64電極表示できる(self, mock_plot: MagicMock, mock_show: MagicMock):
-        self.data.showAll()
+        self.fm.showAll()
         for ch in range(1, 65):
             x_actual, y_actual = mock_plot.call_args_list[ch - 1][0]
             # numpyの配列比較を行う
@@ -29,7 +31,7 @@ class MyTestCase(unittest.TestCase):
     @patch("matplotlib.pyplot.plot")
     def test_1電極表示できる(self, mock_plot: MagicMock, mock_show: MagicMock):
         for ch in range(1, 65):
-            self.data.showSingle(ch)
+            self.fm.showSingle(ch)
 
             x_actual, y_actual = mock_plot.call_args_list[ch - 1][0]
             # numpyの配列比較を行う
